A beam is lit by the bottom bit of a coordinate, per zone
LASERBEAM: 72 items on the disc and no trigger volume reaches one. That
looked like GLASS's problem and it is not. Nothing is supposed to reach
one.

The constructor (0x8002E718) runs at every zone load and takes word 0 of
each endpoint from the OTHER buffer - the zone's own Events chunk, the
#56 rebase. The exec (0x8002E694) tests bit 0 of that word and appends
{raiser, item} to a 32-entry list; a per-frame walk (0x8002EE38)
re-submits the whole list for ever. So JAIL2's corridor grid is X=7352
in COMMON and 7353 in zones 1 and 2: the same coordinate with the bottom
bit set, one unit wide of nothing. 71 of the 72 beams are lit in at
least one zone, none in every zone, and the one lit nowhere is a
(0,0,0)->(0,0,0) dead entry.

Two operand labels were wrong. +18 is not an OBJSLOT - the constructor
overwrites it with q2_coll_find_node(PrimaryColl, &origin_a), so it is
the beam's area. +34 is not a counter - it is the laser kind, clamped to
the dispatcher's six, and all 72 disc beams are already in range. That
also removes a lead #85 was following: the field is not an object handle
at all.

And wiring it found a second bug by counter. Eleven beams queued, 970
faces emitted, seven pixels different on screen. effect.c sorted on
(depth >> 2) while world.c and modeldraw.c had both moved to the
viewport's far_z: against the console's 217-bucket table a beam 700
units away landed in bucket 41 and the wall at the same 700 units in
193, and bucket 0 is drawn first. Every particle, glint and beam more
than a room away was emitted correctly, sorted behind the walls, and
painted over. The test pins it by comparison, not by constant.

Also: --at/--yaw to stand somewhere specific for capture, --no-lasers
for the before picture, and a per-zone LASERBEAM census in the
inspector.

+++ src/game/effect.h
@@ -997,9 +997,15 @@ typedef struct q2_fx_laser_result {
 /* Four groups per lit end. 0x8004908C / 0x80049138 both loop `s2 < 4`. */
 #define Q2_FX_LASER_END_GROUPS 4
 
+/*
+ * `area` is s16 and not u8 because the original's is: the script's own beams
+ * reach here with a raw collision-node index (0x8002EEB4 loads it with `lh`),
+ * and a level with more than 256 nodes would fold two rooms onto one otherwise.
+ * The end bursts still take a byte, which is the original's own narrowing.
+ */
 bool q2_fx_laser(q2_fx_world *w, q2_rng *rng, u32 kind,
                  const s32 from[3], const s32 to[3],
-                 u8 area, u32 ends, q2_fx_laser_result *out);
+                 s16 area, u32 ends, q2_fx_laser_result *out);
 
 /* ------------------------------------------------------------------------- */
 /* Drawing                                                                    */
